requirements.txt
Flask
flask_sqlalchemy
Flask-Cors
waitress
werkzeug
pyjwt
gdown
numpy
pandas
opencv-python-headless
tensorflow
